Adding support for running cpplint to enforce C++ code style compliance.  The violations will display as warnings in VS in release build.
Also added the binaries for python 2.7.9 as cpplint is a python file and we run the build via an exec task in msbuild.

Added the DanceStudio.Targets file to the C++ and C++/CLI core projects.

Added some code to a new HelpController that will be used to execute help operations for the user to learn about Dance Studio.

Updated .gitignore to allow files in a build folder to be considered for addtion.

Added typedefs for the basic C++ types (int, long, etc.).

Added a stdafx.h file to Core but haven't set it up to precompile yet.